Extract all rules from Rule to their own files for a better file structure.
