+++ src/entities/rules/hasMissingFields.ts
@@ -1,0 +1,16 @@
+import { Rule } from '../rule';
+
+export const hasMissingFields = () => new Rule(
+    2,
+    'Missing info',
+    'Fill in name, description, contact info, power need and sound amplification please.',
+    (entity) => {
+        return {
+            triggered: !entity.name ||
+                !entity.description ||
+                !entity.contactInfo ||
+                entity.powerNeed === -1 ||
+                entity.amplifiedSound === -1,
+        };
+    }
+);
