+++ src/entities/rules/isCalculatedAreaTooBig.ts
@@ -1,0 +1,13 @@
+import {
+    MAX_CLUSTER_SIZE
+} from '../../../SETTINGS';
+import { Rule } from '../rule';
+
+export const isCalculatedAreaTooBig = () => new Rule(
+    3,
+    'Too many ppl/vehicles!',
+    'Calculated area need is bigger than the maximum allowed area size! Make another area to fix this.',
+    (entity) => {
+        return { triggered: entity.calculatedAreaNeeded > MAX_CLUSTER_SIZE };
+    }
+);
